Setting collation is broken. Maybe this is the fix?
Hi! Let me preface this by saying I am not a software developer and I do not know too much about PHP, databases etc.
I am running a Bluespice Mediawiki with DPL included. I am trying to order a list with a proper collation for German language but it's not working. I am not the only one with this issue, see https://www.mediawiki.org/wiki/Extension_talk:DynamicPageList3#ordercollation_has_no_effect .
I tried to understand how the code works and why it is broken and I think I might have found the problem: The ordercollation function querys the database to "SHOW CHARACTER SET" which only outputs the default collation for the character set. Only if the collation you want to set is the default collation, it can be set. (Which doesn't make much sense.) Instead, it should be possible to set every supported collation. These are listed with the query "SHOW COLLATION" in the column "Collation".
I might be completely wrong but that's what it looks like to me. Unfortunately, I can't test this since I don't have admin access and my admin has already left for the weekend. :)